[FIX] website_hr_recruitment: fix apply fields
Previously, it wasn't possible to add fields to the form
using the web editor on the apply page.
This commit also set the "job_details" field on the hr.job
model as a translatable field.

Related commit: 0a8c3e745127374a8bf53f3b5931384c0a96978f

task-2990154

X-original-commit: af4c7188a6928a033d0a4ef30f759049d8c59a8c
